Correcting case when no key is found.
When no key is found=>updtae of the file ignored.

+++ grobid-core/src/main/java/org/grobid/core/utilities/GrobidProperties.java
@@ -922,12 +922,13 @@ public class GrobidProperties {
 		}
 		reader.close();
 
-		String newContent = content.replaceAll(lineToReplace, pKey + "="
-				+ pValue);
-
-		FileWriter writer = new FileWriter(pPropertyFile.getAbsoluteFile());
-		writer.write(newContent);
-		writer.close();
+		if (!StringUtils.EMPTY.equals(lineToReplace)) {
+			String newContent = content.replaceAll(lineToReplace, pKey + "="
+					+ pValue);
+			FileWriter writer = new FileWriter(pPropertyFile.getAbsoluteFile());
+			writer.write(newContent);
+			writer.close();
+		}
 	}
 
 	/**
